feat(infrastructure): 重构终端指标服务并优化协议版本指标
- 在 ShardedConnectionManager 中添加 getProtocolVersionConnections 和 getProtocolVersionConnectionCount 方法
- 重构 TerminalMetricsService 为 DeviceMetricsService- 优化协议版本指标注册逻辑，仅注册实际使用的版本
- 合并系统核心指标、线程池指标、分片指标和 EventLoop 指标
- 优化指标计算方法，提高性能和可维护性
